python-backend/requirements.txt
# Startup Analyzer - Django + LangGraph Backend
# Python 3.10+ required

django>=5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langgraph>=0.0.20
python-dotenv>=1.0.0
gunicorn>=21.0.0

# Optional: For Groq instead of OpenAI
# langchain-groq>=0.0.1

# Optional: For RAG features
# chromadb>=0.4.0
# sentence-transformers>=2.2.0
# pypdf>=3.17.0
